+++ CHANGELOG.md
@@ -16,6 +16,10 @@ format of this file follows recommendations from
 - Add new iterm generator directives for changing cursor color and shape
 - Add new iterm generator directive to change the iterm profile
 - Add new iterm generator directive to change the tab or window title background color
+- Add new {env:HOME} interpolation for shell environment variables
+
+### Changed
+
 - Simplify directives in environment_variables generator